requirements.txt
terraform-compliance
invoke
requests
PyGithub